refactor(editor): Split SelectedToolsButton into its own component, renaming
