+++ packages/frontend/editor-ui/src/features/ai/chatHub/components/ChatConversationHeader.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import type { CredentialsMap } from '@/features/ai/chatHub/chat.types';
-import ModelSelector from '@/features/ai/chatHub/components/ModelSelector.vue';
+import ModelSelectorModal from '@/features/ai/chatHub/components/ModelSelectorModal.vue';
 import { useChatHubSidebarState } from '@/features/ai/chatHub/composables/useChatHubSidebarState';
 import { CHAT_VIEW } from '@/features/ai/chatHub/constants';
 import type { ChatHubProvider, ChatModelDto, ChatSessionId } from '@n8n/api-types';
@@ -63,7 +63,7 @@ defineExpose({
 				icon-size="large"
 				@click="onNewChat"
 			/>
-			<ModelSelector
+			<ModelSelectorModal
 				v-if="readyToShowModelSelector"
 				ref="modelSelectorRef"
 				:selected-agent="selectedModel"
