+++ packages/frontend/editor-ui/src/features/ai/chatHub/components/ChatPrompt.vue
@@ -1,13 +1,12 @@
 <script setup lang="ts">
 import { useToast } from '@/app/composables/useToast';
-import NodeIcon from '@/app/components/NodeIcon.vue';
 import { providerDisplayNames } from '@/features/ai/chatHub/constants';
-import { useNodeTypesStore } from '@/app/stores/nodeTypes.store';
 import type { ChatHubLLMProvider, ChatModelDto } from '@n8n/api-types';
-import { N8nButton, N8nIcon, N8nIconButton, N8nInput, N8nText } from '@n8n/design-system';
+import { N8nIconButton, N8nInput, N8nText } from '@n8n/design-system';
 import { useSpeechRecognition } from '@vueuse/core';
 import type { INode } from 'n8n-workflow';
-import { computed, onMounted, ref, useTemplateRef, watch } from 'vue';
+import { computed, ref, useTemplateRef, watch } from 'vue';
+import SelectedToolsButton from './SelectedToolsButton.vue';
 
 const { selectedModel, selectedTools, isMissingCredentials } = defineProps<{
 	isResponding: boolean;
@@ -28,7 +27,6 @@ const emit = defineEmits<{
 const inputRef = useTemplateRef<HTMLElement>('inputRef');
 const message = ref('');
 
-const nodeTypesStore = useNodeTypesStore();
 const toast = useToast();
 
 const speechInput = useSpeechRecognition({
@@ -105,24 +103,6 @@ function onSelectTools() {
 	emit('selectTools');
 }
 
-onMounted(async () => {
-	await nodeTypesStore.loadNodeTypesIfNotLoaded();
-});
-
-const toolCount = computed(() => selectedTools?.length ?? 0);
-
-const displayToolNodeTypes = computed(() => {
-	const tools = selectedTools ?? [];
-	return tools
-		.slice(0, 3)
-		.map((t) => nodeTypesStore.getNodeType(t.type))
-		.filter(Boolean);
-});
-
-const toolsLabel = computed(() =>
-	toolCount.value > 0 ? `${toolCount.value} Tool${toolCount.value > 1 ? 's' : ''}` : 'Tools',
-);
-
 defineExpose({
 	focus: () => inputRef.value?.focus(),
 	setText: (text: string) => {
@@ -170,29 +150,11 @@ defineExpose({
 			/>
 
 			<div :class="$style.tools">
-				<N8nButton
-					:class="$style.toolsButton"
+				<SelectedToolsButton
+					:selected="selectedTools ?? []"
 					:disabled="isMissingCredentials || !selectedModel || isResponding"
-					aria-label="Select tools"
 					@click="onSelectTools"
-				>
-					<span v-if="toolCount" :class="$style.iconStack" aria-hidden="true">
-						<NodeIcon
-							v-for="(nodeType, i) in displayToolNodeTypes"
-							:key="`${nodeType?.name}-${i}`"
-							:style="{ zIndex: displayToolNodeTypes.length - i }"
-							:node-type="nodeType"
-							:class="[$style.icon, { [$style.iconOverlap]: i !== 0 }]"
-							:circle="true"
-							:size="12"
-						/>
-					</span>
-					<span v-else :class="$style.iconFallback" aria-hidden="true">
-						<N8nIcon icon="plus" :size="12" />
-					</span>
-
-					<N8nText size="small" bold>{{ toolsLabel }}</N8nText>
-				</N8nButton>
+				/>
 			</div>
 
 			<div :class="$style.actions">
@@ -291,47 +253,6 @@ defineExpose({
 	display: flex;
 	align-items: center;
 	gap: var(--spacing--2xs);
-}
-
-.toolsButton {
-	display: flex;
-	align-items: center;
-	gap: var(--spacing--2xs);
-	padding: var(--spacing--3xs) var(--spacing--xs);
-	color: var(--color--text);
-	cursor: pointer;
-
-	border-radius: var(--radius);
-	border: var(--border);
-	background: var(--color--background--light-3);
-
-	&:disabled {
-		opacity: 0.6;
-		cursor: not-allowed;
-	}
-}
-
-.iconStack {
-	display: flex;
-	align-items: center;
-	position: relative;
-}
-
-.icon {
-	padding: var(--spacing--4xs);
-	background-color: var(--button--color--background--secondary);
-	border-radius: 50%;
-	outline: 2px var(--color--background--light-3) solid;
-}
-
-.iconOverlap {
-	margin-left: -6px;
-}
-
-.iconFallback {
-	display: flex;
-	align-items: center;
-	justify-content: center;
 }
 
 /* Right-side actions */
