+++ packages/frontend/editor-ui/src/features/ai/chatHub/components/ModelSelectorModal.vue
