+++ packages/frontend/editor-ui/src/features/ai/chatHub/components/ToolsSelectorModal.vue
